Preserve per-dimension bounds in IntermediateType::Array (#675)
* Preserve per-dimension bounds in IntermediateType::Array

Replace the flat `size: Option<u32>` in `IntermediateType::Array` with
`dimensions: Vec<ArrayDimension>` so the codegen can recover lower/upper
bounds for each dimension. This is pre-work PR 0a from the array codegen
plan.

Key changes:
- Add `ArrayDimension` struct with `lower: i32` and `upper: i32`
- Add `array_total_elements()` method to derive total size from dimensions
- Widen `size_in_bytes()` return type from `Option<u8>` to `Option<u32>`
  to support arrays that exceed 255 bytes
- Update construction site in intermediates/array.rs to build dimensions
  from subranges via new `signed_integer_to_i32` helper
- Remove now-unused `calculate_array_size()` function
- Update all match sites and tests across the analyzer crate


* Fix Array field name in resolve_array_type test

Update test added by PR 0f to use the new `dimensions` field
instead of the removed `size` field.


---------

+++ compiler/analyzer/src/intermediates/stdlib_function_block.rs
@@ -65,7 +65,7 @@ fn build_field(
     };
 
     // Calculate size
-    let size = field_type.size_in_bytes().unwrap_or(0) as u32;
+    let size = field_type.size_in_bytes().unwrap_or(0);
 
     let field = IntermediateStructField {
         name: Id::from(name),
